Clean up duplicated test files
