--- archive/duplicates/archive/legacy_builds/build_portable/tests/CMakeFiles/test_performance.dir/flags 2.make
@@ -1,12 +1,0 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 4.0
-
-# compile C with /usr/bin/clang
-C_DEFINES = 
-
-C_INCLUDES = -I/Users/eirikr/Documents/GitHub/bcpl-compiler/src/include
-
-C_FLAGSx86_64 = -std=c17 -target x86_64-apple-macos11 -g -arch x86_64
-
-C_FLAGS = -std=c17 -target x86_64-apple-macos11 -g -arch x86_64
-
