--- archive/duplicates/archive/legacy_builds/build_portable/tests/CMakeFiles/test_platform_compatibility.dir/link 2.txt
@@ -1,1 +1,0 @@
-/usr/bin/clang -std=c17 -target x86_64-apple-macos11 -g -arch x86_64 -Wl,-search_paths_first -Wl,-headerpad_max_install_names CMakeFiles/test_platform_compatibility.dir/test_cross_platform.c.o CMakeFiles/test_platform_compatibility.dir/test_architecture.c.o -o test_platform_compatibility  ../src/libbcpl_runtime.a
