--- archive/duplicates/archive/legacy_builds/build_portable/tests/CMakeFiles/test_runtime_modernization.dir/link 2.txt
@@ -1,1 +1,0 @@
-/usr/bin/clang -std=c17 -target x86_64-apple-macos11 -g -arch x86_64 -Wl,-search_paths_first -Wl,-headerpad_max_install_names CMakeFiles/test_runtime_modernization.dir/test_runtime.c.o CMakeFiles/test_runtime_modernization.dir/test_platform_abstraction.c.o CMakeFiles/test_runtime_modernization.dir/test_memory_safety.c.o -o test_runtime_modernization  ../src/libbcpl_runtime.a
