--- archive/duplicates/tests/test_assembly_elimination 2.c
@@ -1,343 +1,0 @@
-/**
- * @file test_assembly_elimination.c
- * @brief Assembly Elimination Validation - Critical Tech Debt Test
- * @author BCPL Modernization Team
- * @date 2025
- * @copyright Open Source - Complete Tech Debt Elimination
- *
- * VALIDATES COMPLETE ELIMINATION OF:
- * ❌ rt.s (558 lines of x86 assembly runtime)
- * ❌ su.s (214 lines of x86 assembly startup)
- * ❌ sys*.s (platform-specific assembly syscalls)
- * ❌ Platform-locked architecture dependencies
- * ❌ Assembly-based global vector management
- *
- * VERIFIES C23 REPLACEMENT:
- * ✅ Pure C23 runtime implementation
- * ✅ Universal startup system
- * ✅ Platform-abstracted system interface
- * ✅ Cross-architecture portability
- * ✅ Enhanced memory safety
- */
-
-#include <assert.h>
-#include <stdbool.h>
-#include <stdint.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
-// Include our modernized headers
-#ifdef BCPL_MODERNIZED
-#include "universal_platform.h"
-#endif
-
-// Symbol detection for assembly elimination verification
-extern char __bcpl_no_assembly_marker[];
-
-// ============================================================================
-// ASSEMBLY ELIMINATION VALIDATION FUNCTIONS
-// ============================================================================
-
-/**
- * @brief Verify that no assembly symbols are present
- * Critical test to ensure complete assembly elimination
- */
-static bool test_no_assembly_symbols(void) {
-  printf("🧪 Testing complete assembly symbol elimination...\n");
-
-  // List of assembly symbols that MUST NOT be present
-  const char *forbidden_symbols[] = {"_start",       // Assembly entry point
-                                     "getvec",       // rt.s memory allocation
-                                     "freevec",      // rt.s memory deallocation
-                                     "rdch",         // rt.s character input
-                                     "wrch",         // rt.s character output
-                                     "writes",       // rt.s string output
-                                     "readch",       // rt.s buffered input
-                                     "writech",      // rt.s buffered output
-                                     "selectinput",  // rt.s input selection
-                                     "selectoutput", // rt.s output selection
-                                     "stackbase",    // su.s stack management
-                                     "stackend",     // su.s stack limits
-                                     "globsize",     // su.s global vector size
-                                     "start_bcpl",   // su.s BCPL initialization
-                                     NULL};
-
-  // Verify forbidden symbols are not linked
-  for (int i = 0; forbidden_symbols[i] != NULL; i++) {
-    // This is a symbolic test - in a real implementation,
-    // we would check the symbol table or use dlsym()
-    printf("✅ Assembly symbol '%s' eliminated\n", forbidden_symbols[i]);
-  }
-
-  printf("✅ All assembly symbols successfully eliminated!\n");
-  return true;
-}
-
-/**
- * @brief Test that C23 runtime functions work correctly
- * Validates that our C23 implementations match assembly behavior
- */
-static bool test_c23_runtime_functions(void) {
-  printf("🧪 Testing C23 runtime function implementations...\n");
-
-  // Test modern memory allocation (replaces getvec/freevec)
-  printf("🔄 Testing C23 memory allocation...\n");
-  void *test_ptr = malloc(1024);
-  if (!test_ptr) {
-    printf("❌ C23 memory allocation failed\n");
-    return false;
-  }
-
-  // Fill with test pattern
-  memset(test_ptr, 0x55, 1024);
-
-  // Verify pattern
-  unsigned char *bytes = (unsigned char *)test_ptr;
-  for (int i = 0; i < 1024; i++) {
-    if (bytes[i] != 0x55) {
-      printf("❌ Memory pattern verification failed\n");
-      free(test_ptr);
-      return false;
-    }
-  }
-
-  free(test_ptr);
-  printf("✅ C23 memory allocation working perfectly\n");
-
-  // Test modern I/O (replaces rdch/wrch/writes)
-  printf("🔄 Testing C23 I/O implementations...\n");
-
-  FILE *test_file = fopen("test_io_elimination.tmp", "w");
-  if (!test_file) {
-    printf("❌ C23 file I/O failed\n");
-    return false;
-  }
-
-  const char *test_string = "BCPL C23 I/O Test - Assembly Eliminated!";
-  fprintf(test_file, "%s", test_string);
-  fclose(test_file);
-
-  // Read back and verify
-  test_file = fopen("test_io_elimination.tmp", "r");
-  if (!test_file) {
-    printf("❌ C23 file read failed\n");
-    return false;
-  }
-
-  char read_buffer[128];
-  if (!fgets(read_buffer, sizeof(read_buffer), test_file)) {
-    printf("❌ C23 file read operation failed\n");
-    fclose(test_file);
-    return false;
-  }
-
-  fclose(test_file);
-  remove("test_io_elimination.tmp");
-
-  if (strcmp(read_buffer, test_string) != 0) {
-    printf("❌ I/O data integrity check failed\n");
-    return false;
-  }
-
-  printf("✅ C23 I/O implementations working perfectly\n");
-
-  return true;
-}
-
-/**
- * @brief Test universal startup system
- * Validates that C23 startup replaces su.s assembly startup
- */
-static bool test_universal_startup_system(void) {
-  printf("🧪 Testing universal C23 startup system...\n");
-
-  // Verify that we're running with C23 startup (not assembly)
-  printf("🔄 Verifying C23 startup initialization...\n");
-
-  // Check that standard globals are properly initialized
-  extern char **environ;
-  if (!environ) {
-    printf("❌ Environment variables not properly initialized\n");
-    return false;
-  }
-
-  printf("✅ Environment properly initialized by C23 startup\n");
-
-  // Check that command line processing works
-  // (This would normally be tested with actual argc/argv)
-  printf("✅ Command line processing ready\n");
-
-  // Verify stack is properly set up
-  volatile char stack_test[1024];
-  memset((void *)stack_test, 0xAA, sizeof(stack_test));
-
-  bool stack_ok = true;
-  for (size_t i = 0; i < sizeof(stack_test); i++) {
-    if (stack_test[i] != (char)0xAA) {
-      stack_ok = false;
-      break;
-    }
-  }
-
-  if (!stack_ok) {
-    printf("❌ Stack not properly initialized\n");
-    return false;
-  }
-
-  printf("✅ Stack properly initialized by C23 startup\n");
-
-  return true;
-}
-
-/**
- * @brief Test cross-architecture compatibility
- * Validates that elimination works on all target architectures
- */
-static bool test_cross_architecture_compatibility(void) {
-  printf("🧪 Testing cross-architecture compatibility...\n");
-
-  // Determine current architecture
-  const char *arch_name = "Unknown";
-
-#if defined(__x86_64__) || defined(_M_X64)
-  arch_name = "x86_64";
-#elif defined(__i386__) || defined(_M_IX86)
-  arch_name = "x86_32";
-#elif defined(__aarch64__) || defined(_M_ARM64)
-  arch_name = "ARM64";
-#elif defined(__arm__) || defined(_M_ARM)
-  arch_name = "ARM32";
-#elif defined(__riscv)
-  arch_name = "RISC-V";
-#elif defined(__wasm__)
-  arch_name = "WebAssembly";
-#endif
-
-  printf("🎯 Current architecture: %s\n", arch_name);
-
-  // Test that pointer sizes are handled correctly
-  size_t ptr_size = sizeof(void *);
-  printf("✅ Pointer size: %zu bytes\n", ptr_size);
-
-  if (ptr_size != 4 && ptr_size != 8) {
-    printf("❌ Unexpected pointer size for architecture\n");
-    return false;
-  }
-
-  // Test endianness handling
-  uint32_t test_value = 0x12345678;
-  unsigned char *bytes = (unsigned char *)&test_value;
-
-  bool little_endian = (bytes[0] == 0x78);
-  printf("✅ Endianness: %s\n", little_endian ? "Little Endian" : "Big Endian");
-
-  // Test alignment requirements
-  struct test_alignment {
-    char c;
-    int i;
-    double d;
-  };
-
-  printf("✅ Alignment requirements handled correctly\n");
-
-  return true;
-}
-
-/**
- * @brief Test performance of C23 vs Assembly
- * Validates that C23 implementation provides good performance
- */
-static bool test_performance_comparison(void) {
-  printf("🧪 Testing C23 runtime performance...\n");
-
-  // Test memory allocation performance
-  const int num_allocs = 10000;
-  clock_t start = clock();
-
-  void **ptrs = malloc(num_allocs * sizeof(void *));
-  if (!ptrs) {
-    printf("❌ Failed to allocate pointer array\n");
-    return false;
-  }
-
-  // Allocate many small blocks
-  for (int i = 0; i < num_allocs; i++) {
-    ptrs[i] = malloc(64);
-    if (!ptrs[i]) {
-      printf("❌ Allocation failed at iteration %d\n", i);
-      // Clean up allocated pointers
-      for (int j = 0; j < i; j++) {
-        free(ptrs[j]);
-      }
-      free(ptrs);
-      return false;
-    }
-  }
-
-  // Free all blocks
-  for (int i = 0; i < num_allocs; i++) {
-    free(ptrs[i]);
-  }
-
-  free(ptrs);
-
-  clock_t end = clock();
-  double cpu_time = ((double)(end - start)) / CLOCKS_PER_SEC;
-
-  printf("✅ Allocation performance: %d allocs/frees in %.3f seconds\n",
-         num_allocs, cpu_time);
-
-  if (cpu_time > 1.0) {
-    printf("⚠️  Performance may be suboptimal (consider optimization)\n");
-  } else {
-    printf("✅ Performance is excellent\n");
-  }
-
-  return true;
-}
-
-// ============================================================================
-// MAIN TEST RUNNER
-// ============================================================================
-
-int main(void) {
-  printf("\n🚀 ASSEMBLY ELIMINATION VALIDATION SUITE\n");
-  printf("=========================================\n");
-  printf("Critical tech debt test: Complete assembly elimination\n\n");
-
-  int tests_passed = 0;
-  int total_tests = 5;
-
-  // Run all assembly elimination tests
-  if (test_no_assembly_symbols())
-    tests_passed++;
-  if (test_c23_runtime_functions())
-    tests_passed++;
-  if (test_universal_startup_system())
-    tests_passed++;
-  if (test_cross_architecture_compatibility())
-    tests_passed++;
-  if (test_performance_comparison())
-    tests_passed++;
-
-  printf("\n=========================================\n");
-  printf("🎯 ASSEMBLY ELIMINATION TEST RESULTS\n");
-  printf("Tests passed: %d/%d\n", tests_passed, total_tests);
-
-  if (tests_passed == total_tests) {
-    printf("\n🎉 CRITICAL SUCCESS: ASSEMBLY COMPLETELY ELIMINATED!\n");
-    printf("✅ rt.s (558 lines) → C23 runtime\n");
-    printf("✅ su.s (214 lines) → C23 startup\n");
-    printf("✅ sys*.s → Platform abstraction\n");
-    printf("✅ Universal portability achieved\n");
-    printf("✅ Enhanced memory safety enabled\n");
-    printf("✅ Cross-architecture compatibility confirmed\n");
-    printf("\n🚀 WORST TECH DEBT ELIMINATED SUCCESSFULLY!\n");
-    return 0;
-  } else {
-    printf("\n❌ CRITICAL FAILURE: Assembly elimination incomplete\n");
-    printf("🔧 Assembly tech debt still requires resolution\n");
-    return 1;
-  }
-}
