--- archive/duplicates/tests/test_cross_platform 2.c
@@ -1,350 +1,0 @@
-/**
- * @file test_cross_platform.c
- * @brief Cross-Platform Build System Validation
- * @author BCPL Modernization Team
- * @date 2025
- * @copyright Open Source - Complete Tech Debt Elimination
- *
- * VALIDATES ELIMINATION OF:
- * ❌ Fragmented Makefile/CMake hybrid chaos
- * ❌ AppleClang vs Clang compatibility nightmares
- * ❌ Platform-specific build script proliferation
- * ❌ Cross-compilation configuration hell
- * ❌ Manual toolchain detection failures
- *
- * VERIFIES UNIFIED BUILD SYSTEM:
- * ✅ Single CMake system for all platforms
- * ✅ Automatic compiler detection and configuration
- * ✅ Universal cross-compilation support
- * ✅ Platform abstraction working correctly
- * ✅ Enhanced error handling and reporting
- */
-
-#include <assert.h>
-#include <stdint.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
-// Platform detection macros
-#ifdef _WIN32
-#define BCPL_PLATFORM_WINDOWS 1
-#include <windows.h>
-#elif defined(__APPLE__)
-#define BCPL_PLATFORM_MACOS 1
-#include <sys/utsname.h>
-#include <unistd.h>
-#elif defined(__linux__)
-#define BCPL_PLATFORM_LINUX 1
-#include <sys/utsname.h>
-#include <unistd.h>
-#elif defined(__FreeBSD__)
-#define BCPL_PLATFORM_FREEBSD 1
-#include <sys/utsname.h>
-#include <unistd.h>
-#elif defined(__EMSCRIPTEN__)
-#define BCPL_PLATFORM_WASM 1
-#else
-#define BCPL_PLATFORM_UNKNOWN 1
-#endif
-
-// ============================================================================
-// CROSS-PLATFORM VALIDATION FUNCTIONS
-// ============================================================================
-
-/**
- * @brief Test platform detection and identification
- * Validates that build system correctly identifies target platform
- */
-static int test_platform_detection(void) {
-  printf("🧪 Testing platform detection...\n");
-
-  const char *platform_name = "Unknown";
-  int detected = 0;
-
-#ifdef BCPL_PLATFORM_MACOS
-  platform_name = "macOS";
-  detected = 1;
-
-  struct utsname sys_info;
-  if (uname(&sys_info) == 0) {
-    printf("✅ macOS System: %s %s\n", sys_info.sysname, sys_info.release);
-    printf("✅ Machine: %s\n", sys_info.machine);
-  }
-
-#elif defined(BCPL_PLATFORM_LINUX)
-  platform_name = "Linux";
-  detected = 1;
-
-  struct utsname sys_info;
-  if (uname(&sys_info) == 0) {
-    printf("✅ Linux System: %s %s\n", sys_info.sysname, sys_info.release);
-    printf("✅ Machine: %s\n", sys_info.machine);
-  }
-
-#elif defined(BCPL_PLATFORM_FREEBSD)
-  platform_name = "FreeBSD";
-  detected = 1;
-
-  struct utsname sys_info;
-  if (uname(&sys_info) == 0) {
-    printf("✅ FreeBSD System: %s %s\n", sys_info.sysname, sys_info.release);
-    printf("✅ Machine: %s\n", sys_info.machine);
-  }
-
-#elif defined(BCPL_PLATFORM_WINDOWS)
-  platform_name = "Windows";
-  detected = 1;
-
-  OSVERSIONINFO os_info;
-  os_info.dwOSVersionInfoSize = sizeof(OSVERSIONINFO);
-  if (GetVersionEx(&os_info)) {
-    printf("✅ Windows Version: %ld.%ld\n", os_info.dwMajorVersion,
-           os_info.dwMinorVersion);
-  }
-
-#elif defined(BCPL_PLATFORM_WASM)
-  platform_name = "WebAssembly";
-  detected = 1;
-  printf("✅ WebAssembly runtime detected\n");
-#endif
-
-  if (!detected) {
-    printf("❌ Platform detection failed\n");
-    return 0;
-  }
-
-  printf("✅ Platform detected: %s\n", platform_name);
-  return 1;
-}
-
-/**
- * @brief Test architecture detection and compatibility
- * Validates universal architecture support
- */
-static int test_architecture_detection(void) {
-  printf("🧪 Testing architecture detection...\n");
-
-  const char *arch_name = "Unknown";
-  int word_size = sizeof(void *);
-
-#if defined(__x86_64__) || defined(_M_X64)
-  arch_name = "x86_64";
-#elif defined(__i386__) || defined(_M_IX86)
-  arch_name = "x86_32";
-#elif defined(__aarch64__) || defined(_M_ARM64)
-  arch_name = "ARM64";
-#elif defined(__arm__) || defined(_M_ARM)
-  arch_name = "ARM32";
-#elif defined(__riscv)
-#if __riscv_xlen == 64
-  arch_name = "RISC-V 64";
-#else
-  arch_name = "RISC-V 32";
-#endif
-#elif defined(__wasm__)
-  arch_name = "WebAssembly";
-#endif
-
-  printf("✅ Architecture: %s\n", arch_name);
-  printf("✅ Word size: %d bytes\n", word_size);
-
-  // Test endianness
-  uint32_t test_val = 0x12345678;
-  unsigned char *bytes = (unsigned char *)&test_val;
-
-  if (bytes[0] == 0x78) {
-    printf("✅ Byte order: Little Endian\n");
-  } else if (bytes[0] == 0x12) {
-    printf("✅ Byte order: Big Endian\n");
-  } else {
-    printf("❌ Unknown byte order\n");
-    return 0;
-  }
-
-  return 1;
-}
-
-/**
- * @brief Test compiler compatibility and C23 features
- * Validates that modern compilers work correctly
- */
-static int test_compiler_compatibility(void) {
-  printf("🧪 Testing compiler compatibility...\n");
-
-  // Test C23 features
-#ifdef __STDC_VERSION__
-  if (__STDC_VERSION__ >= 202311L) {
-    printf("✅ C23 standard supported\n");
-  } else if (__STDC_VERSION__ >= 201710L) {
-    printf("✅ C17 standard supported\n");
-  } else if (__STDC_VERSION__ >= 201112L) {
-    printf("⚠️  C11 standard (C23 recommended)\n");
-  } else {
-    printf("❌ Outdated C standard\n");
-    return 0;
-  }
-#else
-  printf("⚠️  C standard version unknown\n");
-#endif
-
-  // Test compiler-specific features
-#ifdef __clang__
-  printf("✅ Clang compiler detected\n");
-#ifdef __apple_build_version__
-  printf("✅ Apple Clang variant\n");
-#endif
-#elif defined(__GNUC__)
-  printf("✅ GCC compiler detected\n");
-  printf("✅ GCC version: %d.%d.%d\n", __GNUC__, __GNUC_MINOR__,
-         __GNUC_PATCHLEVEL__);
-#elif defined(_MSC_VER)
-  printf("✅ MSVC compiler detected\n");
-  printf("✅ MSVC version: %d\n", _MSC_VER);
-#else
-  printf("⚠️  Unknown compiler\n");
-#endif
-
-  // Test that enhanced safety features work
-  printf("🔄 Testing enhanced safety features...\n");
-
-  // Stack protection test
-  volatile char buffer[16];
-  memset((void *)buffer, 0xAA, sizeof(buffer));
-
-  // This should not cause issues with stack protection
-  for (int i = 0; i < 16; i++) {
-    if (buffer[i] != (char)0xAA) {
-      printf("❌ Stack protection or memory safety issue\n");
-      return 0;
-    }
-  }
-
-  printf("✅ Stack protection working\n");
-
-  return 1;
-}
-
-/**
- * @brief Test build system integration
- * Validates that CMake configuration works correctly
- */
-static int test_build_system_integration(void) {
-  printf("🧪 Testing build system integration...\n");
-
-  // Test that build definitions are correctly set
-#ifdef BCPL_MODERNIZED
-  printf("✅ BCPL_MODERNIZED flag set\n");
-#else
-  printf("❌ BCPL_MODERNIZED flag missing\n");
-  return 0;
-#endif
-
-#ifdef BCPL_C23_RUNTIME
-  printf("✅ BCPL_C23_RUNTIME flag set\n");
-#else
-  printf("❌ BCPL_C23_RUNTIME flag missing\n");
-  return 0;
-#endif
-
-#ifdef BCPL_NO_ASSEMBLY
-  printf("✅ BCPL_NO_ASSEMBLY flag set\n");
-#else
-  printf("❌ BCPL_NO_ASSEMBLY flag missing\n");
-  return 0;
-#endif
-
-#ifdef BCPL_UNIVERSAL_PLATFORM
-  printf("✅ BCPL_UNIVERSAL_PLATFORM flag set\n");
-#else
-  printf("❌ BCPL_UNIVERSAL_PLATFORM flag missing\n");
-  return 0;
-#endif
-
-  // Test that optimization flags are working
-  int optimization_level = 0;
-
-#ifdef __OPTIMIZE__
-  optimization_level = 1;
-#endif
-
-#ifdef __OPTIMIZE_SIZE__
-  optimization_level = 2;
-#endif
-
-  printf("✅ Optimization level: %d\n", optimization_level);
-
-  return 1;
-}
-
-/**
- * @brief Test cross-compilation capabilities
- * Validates that build system supports cross-compilation
- */
-static int test_cross_compilation_support(void) {
-  printf("🧪 Testing cross-compilation support...\n");
-
-  // Test that we can detect cross-compilation scenarios
-  int is_cross_compiling = 0;
-
-#ifdef CMAKE_CROSSCOMPILING
-  is_cross_compiling = 1;
-  printf("✅ Cross-compilation detected\n");
-#endif
-
-  if (!is_cross_compiling) {
-    printf("✅ Native compilation (cross-compilation ready)\n");
-  }
-
-  // Test that target architecture is properly configured
-  printf("✅ Target configuration validated\n");
-
-  // Test that toolchain detection works
-  printf("✅ Toolchain detection functional\n");
-
-  return 1;
-}
-
-// ============================================================================
-// MAIN TEST RUNNER
-// ============================================================================
-
-int main(void) {
-  printf("\n🚀 CROSS-PLATFORM BUILD SYSTEM VALIDATION\n");
-  printf("==========================================\n");
-  printf("Testing elimination of build system tech debt...\n\n");
-
-  int tests_passed = 0;
-  int total_tests = 5;
-
-  // Run all cross-platform tests
-  if (test_platform_detection())
-    tests_passed++;
-  if (test_architecture_detection())
-    tests_passed++;
-  if (test_compiler_compatibility())
-    tests_passed++;
-  if (test_build_system_integration())
-    tests_passed++;
-  if (test_cross_compilation_support())
-    tests_passed++;
-
-  printf("\n==========================================\n");
-  printf("🎯 CROSS-PLATFORM TEST RESULTS\n");
-  printf("Tests passed: %d/%d\n", tests_passed, total_tests);
-
-  if (tests_passed == total_tests) {
-    printf("\n🎉 BUILD SYSTEM MODERNIZATION SUCCESSFUL!\n");
-    printf("✅ Unified CMake system eliminates build chaos\n");
-    printf("✅ Universal compiler compatibility achieved\n");
-    printf("✅ Cross-platform portability confirmed\n");
-    printf("✅ Enhanced safety and optimization enabled\n");
-    printf("✅ Cross-compilation support validated\n");
-    printf("\n🚀 BUILD SYSTEM TECH DEBT ELIMINATED!\n");
-    return 0;
-  } else {
-    printf("\n❌ Build system modernization incomplete\n");
-    printf("🔧 Build system tech debt requires resolution\n");
-    return 1;
-  }
-}
