--- archive/duplicates/tests/test_memory_safety 2.c
@@ -1,55 +1,0 @@
-/**
- * @file test_memory_safety.c
- * @brief Memory Safety Validation for Modernized BCPL Runtime
- * @author BCPL Modernization Team
- * @date 2025
- */
-
-#include <assert.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
-// Test framework
-#define TEST_ASSERT(condition, message) \
-    do { \
-        if (!(condition)) { \
-            printf("FAIL: %s\n", message); \
-            return 1; \
-        } else { \
-            printf("PASS: %s\n", message); \
-        } \
-    } while(0)
-
-int test_bounds_checking() {
-    TEST_ASSERT(1, "Bounds checking validation");
-    return 0;
-}
-
-int test_memory_leak_detection() {
-    TEST_ASSERT(1, "Memory leak detection");
-    return 0;
-}
-
-int test_buffer_overflow_protection() {
-    TEST_ASSERT(1, "Buffer overflow protection");
-    return 0;
-}
-
-int main() {
-    printf("Memory Safety Tests for BCPL Runtime\n");
-    printf("====================================\n");
-    
-    int result = 0;
-    result |= test_bounds_checking();
-    result |= test_memory_leak_detection();
-    result |= test_buffer_overflow_protection();
-    
-    if (result == 0) {
-        printf("All memory safety tests PASSED\n");
-    } else {
-        printf("Some memory safety tests FAILED\n");
-    }
-    
-    return result;
-}
