--- archive/duplicates/tests/test_platform_abstraction 2.c
@@ -1,313 +1,0 @@
-/**
- * @file test_platform_abstraction.c
- * @brief Platform Abstraction Layer Tests - Tech Debt Elimination Validation
- * @author BCPL Modernization Team
- * @date 2025
- * @copyright Open Source - Complete Tech Debt Elimination
- *
- * VALIDATES COMPLETE ELIMINATION OF:
- * ✅ Platform-specific assembly code dependencies
- * ✅ Architecture-locked build systems
- * ✅ Cross-compilation nightmares
- * ✅ OS-specific system call interfaces
- *
- * VERIFIES UNIVERSAL FEATURES:
- * ✅ Works on ARM64, x86_64, RISC-V architectures
- * ✅ Unified platform interface across macOS, Linux, Windows
- * ✅ WebAssembly compatibility for browser deployment
- * ✅ Thread-safe cross-platform operations
- */
-
-#include <assert.h>
-#include <stdatomic.h>
-#include <stdint.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <time.h>
-
-// Include our modernized platform abstraction
-#ifdef BCPL_MODERNIZED
-#include "universal_platform.h"
-#else
-#include "platform.h"
-#endif
-
-// ============================================================================
-// PLATFORM ABSTRACTION VALIDATION TESTS
-// ============================================================================
-
-/**
- * @brief Test universal memory allocation across all platforms
- * Validates that memory management works identically on all architectures
- */
-static int test_universal_memory_allocation(void) {
-  printf("🧪 Testing universal memory allocation...\n");
-
-  // Test various allocation sizes
-  size_t test_sizes[] = {1, 16, 1024, 65536, 1048576};
-  size_t num_tests = sizeof(test_sizes) / sizeof(test_sizes[0]);
-
-  for (size_t i = 0; i < num_tests; i++) {
-    void *ptr = bcpl_platform_alloc(test_sizes[i]);
-    if (!ptr) {
-      printf("❌ Failed to allocate %zu bytes\n", test_sizes[i]);
-      return 0;
-    }
-
-    // Write test pattern
-    memset(ptr, 0xAA, test_sizes[i]);
-
-    // Verify pattern
-    unsigned char *bytes = (unsigned char *)ptr;
-    for (size_t j = 0; j < test_sizes[i]; j++) {
-      if (bytes[j] != 0xAA) {
-        printf("❌ Memory corruption at offset %zu\n", j);
-        bcpl_platform_free(ptr);
-        return 0;
-      }
-    }
-
-    bcpl_platform_free(ptr);
-    printf("✅ Memory allocation/deallocation test passed for %zu bytes\n",
-           test_sizes[i]);
-  }
-
-  return 1;
-}
-
-/**
- * @brief Test cross-platform file I/O operations
- * Validates that file operations work identically across all platforms
- */
-static int test_universal_file_operations(void) {
-  printf("🧪 Testing universal file I/O...\n");
-
-  const char *test_filename = "test_platform_file.tmp";
-  const char *test_data =
-      "BCPL Modernization Test Data - Universal Platform Support";
-
-  // Test file creation and writing
-  FILE *file = bcpl_platform_fopen(test_filename, "w");
-  if (!file) {
-    printf("❌ Failed to create test file\n");
-    return 0;
-  }
-
-  size_t written = fwrite(test_data, 1, strlen(test_data), file);
-  if (written != strlen(test_data)) {
-    printf("❌ Failed to write complete test data\n");
-    fclose(file);
-    return 0;
-  }
-
-  fclose(file);
-  printf("✅ File creation and writing successful\n");
-
-  // Test file reading
-  file = bcpl_platform_fopen(test_filename, "r");
-  if (!file) {
-    printf("❌ Failed to open test file for reading\n");
-    return 0;
-  }
-
-  char read_buffer[256];
-  size_t read_bytes = fread(read_buffer, 1, sizeof(read_buffer) - 1, file);
-  read_buffer[read_bytes] = '\0';
-
-  if (strcmp(read_buffer, test_data) != 0) {
-    printf("❌ Read data doesn't match written data\n");
-    fclose(file);
-    return 0;
-  }
-
-  fclose(file);
-  printf("✅ File reading and data integrity verified\n");
-
-  // Clean up
-  bcpl_platform_remove(test_filename);
-
-  return 1;
-}
-
-/**
- * @brief Test cross-platform time and timing operations
- * Validates consistent timing across all platforms
- */
-static int test_universal_timing(void) {
-  printf("🧪 Testing universal timing operations...\n");
-
-  // Test high-resolution timing
-  uint64_t start_time = bcpl_platform_get_time_ns();
-
-  // Perform some work
-  volatile int sum = 0;
-  for (int i = 0; i < 1000000; i++) {
-    sum += i;
-  }
-
-  uint64_t end_time = bcpl_platform_get_time_ns();
-  uint64_t elapsed_ns = end_time - start_time;
-
-  if (elapsed_ns == 0) {
-    printf("❌ Timer resolution too low or timer not working\n");
-    return 0;
-  }
-
-  printf("✅ High-resolution timing: %llu nanoseconds elapsed\n",
-         (unsigned long long)elapsed_ns);
-
-  // Test sleep functionality
-  start_time = bcpl_platform_get_time_ns();
-  bcpl_platform_sleep_ms(10); // Sleep for 10ms
-  end_time = bcpl_platform_get_time_ns();
-
-  uint64_t sleep_elapsed = (end_time - start_time) / 1000000; // Convert to ms
-
-  if (sleep_elapsed < 8 || sleep_elapsed > 50) {
-    printf("❌ Sleep timing inaccurate: expected ~10ms, got %llums\n",
-           (unsigned long long)sleep_elapsed);
-    return 0;
-  }
-
-  printf("✅ Sleep timing accurate: %llums\n",
-         (unsigned long long)sleep_elapsed);
-
-  return 1;
-}
-
-/**
- * @brief Test architecture-specific optimizations
- * Validates that optimizations work correctly on current architecture
- */
-static int test_architecture_optimizations(void) {
-  printf("🧪 Testing architecture-specific optimizations...\n");
-
-  // Test CPU feature detection
-  bcpl_cpu_features_t features = bcpl_platform_get_cpu_features();
-
-  printf("✅ CPU Architecture: %s\n", features.arch_name);
-  printf("✅ CPU Cores: %d\n", features.core_count);
-  printf("✅ SIMD Support: %s\n", features.has_simd ? "Yes" : "No");
-  printf("✅ AES Support: %s\n", features.has_aes ? "Yes" : "No");
-
-  // Test optimized memory operations
-  const size_t test_size = 8192;
-  void *src = bcpl_platform_alloc(test_size);
-  void *dst = bcpl_platform_alloc(test_size);
-
-  if (!src || !dst) {
-    printf("❌ Failed to allocate test buffers\n");
-    bcpl_platform_free(src);
-    bcpl_platform_free(dst);
-    return 0;
-  }
-
-  // Fill source with test pattern
-  for (size_t i = 0; i < test_size; i++) {
-    ((char *)src)[i] = (char)(i & 0xFF);
-  }
-
-  // Test optimized copy
-  uint64_t start = bcpl_platform_get_time_ns();
-  bcpl_platform_memcpy_optimized(dst, src, test_size);
-  uint64_t end = bcpl_platform_get_time_ns();
-
-  // Verify copy
-  if (memcmp(src, dst, test_size) != 0) {
-    printf("❌ Optimized memory copy failed\n");
-    bcpl_platform_free(src);
-    bcpl_platform_free(dst);
-    return 0;
-  }
-
-  printf("✅ Optimized memory copy: %llu ns for %zu bytes\n",
-         (unsigned long long)(end - start), test_size);
-
-  bcpl_platform_free(src);
-  bcpl_platform_free(dst);
-
-  return 1;
-}
-
-/**
- * @brief Test thread safety of platform operations
- * Validates that platform abstraction is thread-safe
- */
-static int test_thread_safety(void) {
-  printf("🧪 Testing thread safety...\n");
-
-  // Test atomic operations
-  atomic_int counter = 0;
-
-  // Simulate concurrent access
-  for (int i = 0; i < 1000; i++) {
-    atomic_fetch_add(&counter, 1);
-  }
-
-  if (atomic_load(&counter) != 1000) {
-    printf("❌ Atomic operations failed: expected 1000, got %d\n",
-           atomic_load(&counter));
-    return 0;
-  }
-
-  printf("✅ Atomic operations working correctly\n");
-
-  // Test thread-safe memory allocation
-  void *ptrs[100];
-  for (int i = 0; i < 100; i++) {
-    ptrs[i] = bcpl_platform_alloc(1024);
-    if (!ptrs[i]) {
-      printf("❌ Thread-safe allocation failed at iteration %d\n", i);
-      return 0;
-    }
-  }
-
-  for (int i = 0; i < 100; i++) {
-    bcpl_platform_free(ptrs[i]);
-  }
-
-  printf("✅ Thread-safe memory allocation working correctly\n");
-
-  return 1;
-}
-
-// ============================================================================
-// MAIN TEST RUNNER
-// ============================================================================
-
-int main(void) {
-  printf("\n🚀 BCPL PLATFORM ABSTRACTION VALIDATION SUITE\n");
-  printf("==============================================\n");
-  printf("Testing complete elimination of platform-specific tech debt...\n\n");
-
-  int tests_passed = 0;
-  int total_tests = 5;
-
-  // Run all platform abstraction tests
-  if (test_universal_memory_allocation())
-    tests_passed++;
-  if (test_universal_file_operations())
-    tests_passed++;
-  if (test_universal_timing())
-    tests_passed++;
-  if (test_architecture_optimizations())
-    tests_passed++;
-  if (test_thread_safety())
-    tests_passed++;
-
-  printf("\n==============================================\n");
-  printf("🎯 PLATFORM ABSTRACTION TEST RESULTS\n");
-  printf("Tests passed: %d/%d\n", tests_passed, total_tests);
-
-  if (tests_passed == total_tests) {
-    printf("✅ ALL PLATFORM ABSTRACTION TESTS PASSED!\n");
-    printf("🚀 Tech debt elimination SUCCESSFUL!\n");
-    printf("✅ Universal portability achieved!\n");
-    return 0;
-  } else {
-    printf("❌ Some platform abstraction tests failed\n");
-    printf("🔧 Platform tech debt still needs resolution\n");
-    return 1;
-  }
-}
