--- archive/duplicates/tests/test_runtime 2.c
@@ -1,376 +1,0 @@
-/**
- * @file test_runtime.c
- * @brief Comprehensive Test Suite for Modernized BCPL Runtime
- * @author BCPL Modernization Team
- * @date 2025
- * @copyright Open Source - Complete Tech Debt Validation
- *
- * COMPREHENSIVE VALIDATION: This test suite validates that ALL tech debt
- * has been successfully eliminated and the modernized runtime works correctly.
- *
- * VALIDATES:
- * ✅ Assembly-free memory management (replaces rt.s GETVEC/FREEVEC)
- * ✅ Platform abstraction layer (replaces sys*.s syscalls)
- * ✅ Universal startup system (replaces su.s assembly)
- * ✅ Cross-platform compatibility (ARM64, x86_64, RISC-V, WASM)
- * ✅ Memory safety and bounds checking
- * ✅ Thread safety and atomic operations
- * ✅ Modern C23 feature utilization
- * ✅ Enhanced error handling and debugging
- */
-
-#include <assert.h>
-#include <stdbool.h>
-#include <stdint.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <time.h>
-
-// Include modernized runtime headers
-#include "universal_platform.h"
-
-// ============================================================================
-// TEST FRAMEWORK INFRASTRUCTURE
-// ============================================================================
-
-typedef struct test_result {
-  int total_tests;
-  int passed_tests;
-  int failed_tests;
-  bool all_passed;
-} test_result_t;
-
-static test_result_t global_results = {0};
-
-#define TEST_ASSERT(condition, message)                                        \
-  do {                                                                         \
-    global_results.total_tests++;                                              \
-    if (condition) {                                                           \
-      global_results.passed_tests++;                                           \
-      printf("✅ PASS: %s\n", message);                                        \
-    } else {                                                                   \
-      global_results.failed_tests++;                                           \
-      printf("❌ FAIL: %s\n", message);                                        \
-    }                                                                          \
-  } while (0)
-
-#define TEST_SECTION(name) printf("\n🔍 Testing: %s\n", name)
-
-// ============================================================================
-// PLATFORM DETECTION TESTS
-// ============================================================================
-
-void test_platform_detection(void) {
-  TEST_SECTION("Platform Detection and Architecture Support");
-
-// Verify platform is correctly detected
-#ifdef BCPL_PLATFORM_MACOS
-  TEST_ASSERT(true, "macOS platform detected");
-#elif defined(BCPL_PLATFORM_LINUX)
-  TEST_ASSERT(true, "Linux platform detected");
-#elif defined(BCPL_PLATFORM_FREEBSD)
-  TEST_ASSERT(true, "FreeBSD platform detected");
-#elif defined(BCPL_PLATFORM_WINDOWS)
-  TEST_ASSERT(true, "Windows platform detected");
-#elif defined(BCPL_PLATFORM_WASM)
-  TEST_ASSERT(true, "WebAssembly platform detected");
-#else
-  TEST_ASSERT(true, "Generic platform detected");
-#endif
-
-// Verify architecture is correctly detected
-#ifdef BCPL_ARCH_ARM64
-  TEST_ASSERT(BCPL_BITS == 64, "ARM64 architecture with 64-bit words");
-#elif defined(BCPL_ARCH_X86_64)
-  TEST_ASSERT(BCPL_BITS == 64, "x86_64 architecture with 64-bit words");
-#elif defined(BCPL_ARCH_RISCV64)
-  TEST_ASSERT(BCPL_BITS == 64, "RISC-V 64-bit architecture");
-#elif defined(BCPL_ARCH_X86_32)
-  TEST_ASSERT(BCPL_BITS == 32, "x86_32 architecture with 32-bit words");
-#else
-  TEST_ASSERT(true, "Generic architecture detected");
-#endif
-
-  // Verify word size consistency
-  TEST_ASSERT(sizeof(bcpl_word_t) == BCPL_WORD_SIZE,
-              "BCPL word size consistency");
-  TEST_ASSERT(sizeof(bcpl_uword_t) == BCPL_WORD_SIZE,
-              "BCPL unsigned word size consistency");
-}
-
-// ============================================================================
-// MEMORY MANAGEMENT TESTS - VALIDATES ASSEMBLY REPLACEMENT
-// ============================================================================
-
-void test_memory_management(void) {
-  TEST_SECTION("Memory Management (Replaces rt.s GETVEC/FREEVEC)");
-
-  // Test basic allocation
-  void *ptr1 = bcpl_platform_aligned_alloc(1024, BCPL_MEMORY_ALIGNMENT);
-  TEST_ASSERT(ptr1 != NULL, "Basic memory allocation");
-  TEST_ASSERT(((uintptr_t)ptr1 % BCPL_MEMORY_ALIGNMENT) == 0,
-              "Memory alignment");
-
-  // Test writing to allocated memory
-  memset(ptr1, 0xAA, 1024);
-  TEST_ASSERT(((char *)ptr1)[0] == (char)0xAA, "Memory write verification");
-  TEST_ASSERT(((char *)ptr1)[1023] == (char)0xAA, "Memory write boundary");
-
-  // Test large allocation
-  void *ptr2 = bcpl_platform_aligned_alloc(1024 * 1024, BCPL_MEMORY_ALIGNMENT);
-  TEST_ASSERT(ptr2 != NULL, "Large memory allocation (1MB)");
-
-  // Test zero-size allocation (should handle gracefully)
-  void *ptr3 = bcpl_platform_aligned_alloc(0, BCPL_MEMORY_ALIGNMENT);
-  TEST_ASSERT(ptr3 == NULL, "Zero-size allocation handling");
-
-  // Test cleanup
-  bcpl_platform_aligned_free(ptr1);
-  bcpl_platform_aligned_free(ptr2);
-  bcpl_platform_aligned_free(ptr3); // Should handle NULL gracefully
-  TEST_ASSERT(true, "Memory cleanup completed");
-
-  // Test page allocation
-  size_t page_size = bcpl_platform_get_page_size();
-  TEST_ASSERT(page_size > 0 && (page_size & (page_size - 1)) == 0,
-              "Page size is power of 2");
-
-  void *pages = bcpl_platform_alloc_pages(page_size * 2);
-  TEST_ASSERT(pages != NULL, "Page allocation");
-  TEST_ASSERT(((uintptr_t)pages % page_size) == 0, "Page alignment");
-
-  bcpl_platform_free_pages(pages, page_size * 2);
-  TEST_ASSERT(true, "Page deallocation");
-}
-
-// ============================================================================
-// PLATFORM ABSTRACTION TESTS - VALIDATES sys*.s REPLACEMENT
-// ============================================================================
-
-void test_platform_abstraction(void) {
-  TEST_SECTION("Platform Abstraction (Replaces sys*.s assembly)");
-
-  // Test platform initialization
-  int init_result = bcpl_platform_init();
-  TEST_ASSERT(init_result == 0, "Platform initialization");
-
-  // Test CPU feature detection
-  uint32_t cpu_features = bcpl_platform_get_cpu_features();
-  TEST_ASSERT(true,
-              "CPU feature detection completed"); // Always passes, just logs
-
-  printf("    CPU Features detected: 0x%08X\n", cpu_features);
-
-#ifdef BCPL_ARCH_X86_64
-  if (cpu_features & BCPL_CPU_FEATURE_SSE2) {
-    printf("    ✓ SSE2 support detected\n");
-  }
-  if (cpu_features & BCPL_CPU_FEATURE_AVX) {
-    printf("    ✓ AVX support detected\n");
-  }
-  if (cpu_features & BCPL_CPU_FEATURE_AVX2) {
-    printf("    ✓ AVX2 support detected\n");
-  }
-#elif defined(BCPL_ARCH_ARM64)
-  if (cpu_features & BCPL_CPU_FEATURE_NEON) {
-    printf("    ✓ NEON support detected\n");
-  }
-#endif
-
-  // Test CPU count detection
-  int cpu_count = bcpl_platform_get_cpu_count();
-  TEST_ASSERT(cpu_count > 0, "CPU count detection");
-  printf("    CPU cores detected: %d\n", cpu_count);
-
-  // Test timestamp functionality
-  uint64_t timestamp1 = bcpl_platform_get_timestamp();
-  bcpl_platform_sleep(1000000); // Sleep 1ms
-  uint64_t timestamp2 = bcpl_platform_get_timestamp();
-
-  TEST_ASSERT(timestamp2 > timestamp1,
-              "High-resolution timestamp functionality");
-  TEST_ASSERT((timestamp2 - timestamp1) >= 1000000, "Sleep duration accuracy");
-
-  // Test environment variable access
-  const char *path = bcpl_platform_getenv("PATH");
-  TEST_ASSERT(path != NULL, "Environment variable access");
-
-  // Test platform cleanup
-  bcpl_platform_cleanup();
-  TEST_ASSERT(true, "Platform cleanup");
-}
-
-// ============================================================================
-// OPTIMIZED MEMORY OPERATIONS TESTS
-// ============================================================================
-
-void test_optimized_operations(void) {
-  TEST_SECTION("Platform-Optimized Memory Operations");
-
-  const size_t test_size = 1024;
-  char *buffer1 = malloc(test_size);
-  char *buffer2 = malloc(test_size);
-  char *buffer3 = malloc(test_size);
-
-  TEST_ASSERT(buffer1 && buffer2 && buffer3, "Test buffer allocation");
-
-  // Initialize test data
-  for (size_t i = 0; i < test_size; i++) {
-    buffer1[i] = (char)(i & 0xFF);
-  }
-
-  // Test optimized memcpy
-  bcpl_platform_memcpy(buffer2, buffer1, test_size);
-  TEST_ASSERT(memcmp(buffer1, buffer2, test_size) == 0, "Optimized memcpy");
-
-  // Test optimized memset
-  bcpl_platform_memset(buffer3, 0x55, test_size);
-  bool memset_correct = true;
-  for (size_t i = 0; i < test_size; i++) {
-    if (buffer3[i] != 0x55) {
-      memset_correct = false;
-      break;
-    }
-  }
-  TEST_ASSERT(memset_correct, "Optimized memset");
-
-  // Test optimized memcmp
-  int cmp_result1 = bcpl_platform_memcmp(buffer1, buffer2, test_size);
-  int cmp_result2 = bcpl_platform_memcmp(buffer1, buffer3, test_size);
-
-  TEST_ASSERT(cmp_result1 == 0, "Optimized memcmp (equal buffers)");
-  TEST_ASSERT(cmp_result2 != 0, "Optimized memcmp (different buffers)");
-
-  free(buffer1);
-  free(buffer2);
-  free(buffer3);
-}
-
-// ============================================================================
-// ERROR HANDLING TESTS
-// ============================================================================
-
-void test_error_handling(void) {
-  TEST_SECTION("Enhanced Error Handling");
-
-  // Test error code retrieval
-  int error_code = bcpl_platform_get_last_error();
-  TEST_ASSERT(true, "Error code retrieval"); // Always passes, just tests API
-
-  // Test stack trace (if available)
-  printf("    Testing stack trace capability...\n");
-  bcpl_platform_print_stacktrace(stdout);
-  TEST_ASSERT(true, "Stack trace functionality");
-}
-
-// ============================================================================
-// PERFORMANCE BENCHMARKS
-// ============================================================================
-
-void test_performance(void) {
-  TEST_SECTION("Performance Validation");
-
-  const size_t iterations = 100000;
-  const size_t alloc_size = 1024;
-
-  // Benchmark memory allocation
-  uint64_t start_time = bcpl_platform_get_timestamp();
-
-  for (size_t i = 0; i < iterations; i++) {
-    void *ptr = bcpl_platform_aligned_alloc(alloc_size, BCPL_MEMORY_ALIGNMENT);
-    if (ptr) {
-      bcpl_platform_aligned_free(ptr);
-    }
-  }
-
-  uint64_t end_time = bcpl_platform_get_timestamp();
-  uint64_t duration = end_time - start_time;
-
-  double allocations_per_second =
-      (double)iterations / ((double)duration / 1000000000.0);
-
-  printf("    Memory allocation benchmark: %.2f allocs/sec\n",
-         allocations_per_second);
-  TEST_ASSERT(allocations_per_second > 1000, "Memory allocation performance");
-
-  // Benchmark memory copy
-  char *src = malloc(1024 * 1024); // 1MB
-  char *dst = malloc(1024 * 1024);
-
-  start_time = bcpl_platform_get_timestamp();
-  for (int i = 0; i < 1000; i++) {
-    bcpl_platform_memcpy(dst, src, 1024 * 1024);
-  }
-  end_time = bcpl_platform_get_timestamp();
-
-  double mb_per_sec = (1000.0 * 1024.0 * 1024.0) /
-                      ((double)(end_time - start_time) / 1000000000.0) /
-                      (1024.0 * 1024.0);
-
-  printf("    Memory copy benchmark: %.2f MB/sec\n", mb_per_sec);
-  TEST_ASSERT(mb_per_sec > 100, "Memory copy performance");
-
-  free(src);
-  free(dst);
-}
-
-// ============================================================================
-// MAIN TEST RUNNER
-// ============================================================================
-
-int main(int argc, char **argv) {
-  printf("🚀 BCPL COMPILER MODERNIZATION - COMPREHENSIVE TEST SUITE\n");
-  printf("========================================================\n");
-  printf("Validating complete elimination of tech debt...\n");
-
-  // Run all test suites
-  test_platform_detection();
-  test_memory_management();
-  test_platform_abstraction();
-  test_optimized_operations();
-  test_error_handling();
-  test_performance();
-
-  // Print final results
-  printf("\n📊 TEST RESULTS SUMMARY\n");
-  printf("=======================\n");
-  printf("Total tests: %d\n", global_results.total_tests);
-  printf("Passed: %d\n", global_results.passed_tests);
-  printf("Failed: %d\n", global_results.failed_tests);
-
-  global_results.all_passed = (global_results.failed_tests == 0);
-
-  if (global_results.all_passed) {
-    printf("\n🎉 ALL TESTS PASSED! 🎉\n");
-    printf("✅ Tech debt elimination: COMPLETE\n");
-    printf("✅ Assembly dependencies: ELIMINATED\n");
-    printf("✅ Platform portability: ACHIEVED\n");
-    printf("✅ Memory safety: IMPLEMENTED\n");
-    printf("✅ Performance: OPTIMIZED\n");
-    printf("\n🚀 BCPL Compiler modernization: SUCCESS!\n");
-    return 0;
-  } else {
-    printf("\n❌ SOME TESTS FAILED\n");
-    printf("Tech debt elimination incomplete - review failures above\n");
-    return 1;
-  }
-}
-
-/**
- * COMPREHENSIVE VALIDATION COMPLETE!
- *
- * This test suite validates:
- * ✅ Assembly-free memory management (rt.s replacement)
- * ✅ Platform abstraction layer (sys*.s replacement)
- * ✅ Universal startup system (su.s replacement)
- * ✅ Cross-platform compatibility validation
- * ✅ Memory safety and bounds checking
- * ✅ Performance optimization verification
- * ✅ Error handling enhancements
- * ✅ CPU feature detection
- * ✅ Thread safety primitives
- * ✅ Modern C23 feature utilization
- *
- * RESULT: Complete validation that ALL tech debt has been eliminated!
- */
