--- archive/legacy_cmake/CMakeLists_modern 2.txt
@@ -1,166 +1,0 @@
-cmake_minimum_required(VERSION 3.20)
-
-# Modern BCPL Compiler Build System
-project(BCPL 
-    VERSION 2025.1.0
-    DESCRIPTION "Modernized BCPL Compiler with C23 Runtime"
-    LANGUAGES C ASM
-)
-
-# Compiler Requirements - Accept AppleClang and Clang
-if(NOT CMAKE_C_COMPILER_ID MATCHES ".*Clang.*")
-    message(FATAL_ERROR "Clang is required to build the modernized BCPL compiler")
-endif()
-
-# C23 Standard with Enhanced Features
-set(CMAKE_C_STANDARD 23)
-set(CMAKE_C_STANDARD_REQUIRED ON)
-set(CMAKE_C_EXTENSIONS OFF)
-
-# Build Configuration
-set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
-
-# =============================================================================
-# PLATFORM AND ARCHITECTURE DETECTION
-# =============================================================================
-
-if(NOT DEFINED BITS)
-    if(CMAKE_SIZEOF_VOID_P EQUAL 8)
-        set(BITS 64)
-    elseif(CMAKE_SIZEOF_VOID_P EQUAL 4)
-        set(BITS 32)
-    else()
-        set(BITS 16)
-    endif()
-endif()
-
-message(STATUS "Building BCPL for ${BITS}-bit architecture on ${CMAKE_SYSTEM_NAME}")
-
-# Platform-specific settings
-if(APPLE)
-    set(BCPL_PLATFORM "MACOS")
-    set(BCPL_PLATFORM_SOURCES platform/macos.c)
-elseif(CMAKE_SYSTEM_NAME STREQUAL "Linux")
-    set(BCPL_PLATFORM "LINUX")
-    set(BCPL_PLATFORM_SOURCES platform/linux.c)
-elseif(CMAKE_SYSTEM_NAME STREQUAL "FreeBSD")
-    set(BCPL_PLATFORM "FREEBSD")
-    set(BCPL_PLATFORM_SOURCES platform/freebsd.c)
-else()
-    set(BCPL_PLATFORM "GENERIC")
-    set(BCPL_PLATFORM_SOURCES platform/generic.c)
-endif()
-
-# Include directories
-include_directories(
-    ${CMAKE_CURRENT_SOURCE_DIR}/include
-    ${CMAKE_CURRENT_SOURCE_DIR}
-)
-
-# =============================================================================
-# COMPILER FLAGS
-# =============================================================================
-
-set(BCPL_C_FLAGS
-    -std=c23
-    -Wall -Wextra -Wpedantic
-    -DBITS=${BITS}
-    -DBCPL_PLATFORM_${BCPL_PLATFORM}=1
-)
-
-if(CMAKE_BUILD_TYPE STREQUAL "Debug")
-    list(APPEND BCPL_C_FLAGS -g3 -O0 -DDEBUG=1)
-else()
-    list(APPEND BCPL_C_FLAGS -O2 -g)
-endif()
-
-# =============================================================================
-# MODERNIZED BCPL RUNTIME LIBRARY
-# =============================================================================
-
-set(BCPL_RUNTIME_SOURCES
-    runtime/bcpl_runtime.c
-    runtime/memory.c
-    runtime/io.c
-    runtime/strings.c
-    runtime/vectors.c
-    ${BCPL_PLATFORM_SOURCES}
-)
-
-# Create the modern runtime library
-add_library(bcpl_runtime STATIC ${BCPL_RUNTIME_SOURCES})
-target_compile_options(bcpl_runtime PRIVATE ${BCPL_C_FLAGS})
-target_include_directories(bcpl_runtime 
-    PUBLIC 
-        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-        $<INSTALL_INTERFACE:include>
-)
-
-# =============================================================================
-# COMPILER BACKEND TOOLS
-# =============================================================================
-
-# Code Generator
-add_executable(cg cg.c oc.c)
-target_compile_options(cg PRIVATE ${BCPL_C_FLAGS})
-target_link_libraries(cg PRIVATE bcpl_runtime)
-
-# Optimizer
-add_executable(op op.c pt.c)
-target_compile_options(op PRIVATE ${BCPL_C_FLAGS})
-target_link_libraries(op PRIVATE bcpl_runtime)
-
-# =============================================================================
-# ENHANCED LLVM BACKEND
-# =============================================================================
-
-find_package(LLVM QUIET CONFIG)
-if(LLVM_FOUND)
-    message(STATUS "Found LLVM ${LLVM_PACKAGE_VERSION}")
-    
-    include_directories(${LLVM_INCLUDE_DIRS})
-    separate_arguments(LLVM_DEFINITIONS_LIST UNIX_COMMAND "${LLVM_DEFINITIONS}")
-    add_definitions(${LLVM_DEFINITIONS_LIST})
-    
-    llvm_map_components_to_libnames(LLCG_LIBS
-        core bitwriter native analysis target
-    )
-    
-    add_executable(llcg llcg.c)
-    target_compile_options(llcg PRIVATE ${BCPL_C_FLAGS})
-    target_link_libraries(llcg PRIVATE ${LLCG_LIBS} bcpl_runtime)
-    
-    set(LLVM_AVAILABLE TRUE)
-else()
-    message(STATUS "LLVM not found - building without LLVM backend")
-    set(LLVM_AVAILABLE FALSE)
-endif()
-
-# =============================================================================
-# INSTALLATION
-# =============================================================================
-
-install(TARGETS cg op bcpl_runtime
-    RUNTIME DESTINATION bin
-    LIBRARY DESTINATION lib
-    ARCHIVE DESTINATION lib
-)
-
-if(LLVM_AVAILABLE)
-    install(TARGETS llcg RUNTIME DESTINATION bin)
-endif()
-
-install(DIRECTORY include/ DESTINATION include/bcpl)
-
-# =============================================================================
-# SUMMARY
-# =============================================================================
-
-message(STATUS "=== BCPL Compiler Build Configuration ===")
-message(STATUS "Build type: ${CMAKE_BUILD_TYPE}")
-message(STATUS "Platform: ${BCPL_PLATFORM}")
-message(STATUS "Architecture: ${BITS}-bit")
-message(STATUS "C Compiler: ${CMAKE_C_COMPILER_ID} ${CMAKE_C_COMPILER_VERSION}")
-message(STATUS "LLVM Backend: ${LLVM_AVAILABLE}")
-message(STATUS "Runtime: Pure C23 implementation")
-message(STATUS "==========================================")
