--- archive/legacy_cmake/CMakeLists_simple 2.txt
@@ -1,13 +1,0 @@
-# Simple test CMakeLists.txt to verify modular structure
-cmake_minimum_required(VERSION 3.20)
-
-# Include centralized options
-list(APPEND CMAKE_MODULE_PATH ${CMAKE_CURRENT_SOURCE_DIR}/../cmake)
-include(Options)
-
-# Create a simple test target to verify the structure works
-add_executable(structure_test structure_test.c)
-apply_tooling_flags(structure_test)
-
-# Register test
-add_test(NAME structure_test COMMAND structure_test)
