+++ tests/test_performance.c
@@ -5,90 +5,87 @@
  * @date 2025
  */
 
-#define _POSIX_C_SOURCE 200809L
 #include <assert.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <string.h>
 #include <time.h>
+#include <string.h>
 
 // Test framework
-#define TEST_ASSERT(condition, message)                                        \
-  do {                                                                         \
-    if (!(condition)) {                                                        \
-      printf("FAIL: %s\n", message);                                           \
-      return 1;                                                                \
-    } else {                                                                   \
-      printf("PASS: %s\n", message);                                           \
-    }                                                                          \
-  } while (0)
+#define TEST_ASSERT(condition, message) \
+    do { \
+        if (!(condition)) { \
+            printf("FAIL: %s\n", message); \
+            return 1; \
+        } else { \
+            printf("PASS: %s\n", message); \
+        } \
+    } while(0)
 
 double get_time_ms() {
-  struct timespec ts;
-  clock_gettime(CLOCK_MONOTONIC, &ts);
-  return ts.tv_sec * 1000.0 + ts.tv_nsec / 1000000.0;
+    struct timespec ts;
+    clock_gettime(CLOCK_MONOTONIC, &ts);
+    return ts.tv_sec * 1000.0 + ts.tv_nsec / 1000000.0;
 }
 
 int test_memory_allocation_performance() {
-  const int iterations = 10000;
-  double start_time = get_time_ms();
-
-  for (int i = 0; i < iterations; i++) {
-    void *ptr = malloc(1024);
-    if (ptr) {
-      memset(ptr, 0, 1024);
-      free(ptr);
+    const int iterations = 10000;
+    double start_time = get_time_ms();
+    
+    for (int i = 0; i < iterations; i++) {
+        void *ptr = malloc(1024);
+        if (ptr) {
+            memset(ptr, 0, 1024);
+            free(ptr);
+        }
     }
-  }
-
-  double end_time = get_time_ms();
-  double duration = end_time - start_time;
-
-  printf("Memory allocation performance: %.2f ms for %d iterations\n", duration,
-         iterations);
-  TEST_ASSERT(duration < 1000.0, "Memory allocation performance is acceptable");
-  return 0;
+    
+    double end_time = get_time_ms();
+    double duration = end_time - start_time;
+    
+    printf("Memory allocation performance: %.2f ms for %d iterations\n", duration, iterations);
+    TEST_ASSERT(duration < 1000.0, "Memory allocation performance is acceptable");
+    return 0;
 }
 
 int test_compilation_speed() {
-  // This would typically measure compilation time
-  // For now, just verify the infrastructure exists
-  TEST_ASSERT(1, "Compilation speed measurement infrastructure");
-  return 0;
+    // This would typically measure compilation time
+    // For now, just verify the infrastructure exists
+    TEST_ASSERT(1, "Compilation speed measurement infrastructure");
+    return 0;
 }
 
 int test_runtime_efficiency() {
-  // Basic runtime efficiency test
-  const int iterations = 100000;
-  volatile int sum = 0;
-
-  double start_time = get_time_ms();
-  for (int i = 0; i < iterations; i++) {
-    sum += i;
-  }
-  double end_time = get_time_ms();
-
-  double duration = end_time - start_time;
-  printf("Runtime efficiency: %.2f ms for %d operations\n", duration,
-         iterations);
-  TEST_ASSERT(duration < 100.0, "Runtime efficiency is acceptable");
-  return 0;
+    // Basic runtime efficiency test
+    const int iterations = 100000;
+    volatile int sum = 0;
+    
+    double start_time = get_time_ms();
+    for (int i = 0; i < iterations; i++) {
+        sum += i;
+    }
+    double end_time = get_time_ms();
+    
+    double duration = end_time - start_time;
+    printf("Runtime efficiency: %.2f ms for %d operations\n", duration, iterations);
+    TEST_ASSERT(duration < 100.0, "Runtime efficiency is acceptable");
+    return 0;
 }
 
 int main() {
-  printf("Performance Benchmarks for BCPL Compiler\n");
-  printf("========================================\n");
-
-  int result = 0;
-  result |= test_memory_allocation_performance();
-  result |= test_compilation_speed();
-  result |= test_runtime_efficiency();
-
-  if (result == 0) {
-    printf("All performance tests PASSED\n");
-  } else {
-    printf("Some performance tests FAILED\n");
-  }
-
-  return result;
+    printf("Performance Benchmarks for BCPL Compiler\n");
+    printf("========================================\n");
+    
+    int result = 0;
+    result |= test_memory_allocation_performance();
+    result |= test_compilation_speed();
+    result |= test_runtime_efficiency();
+    
+    if (result == 0) {
+        printf("All performance tests PASSED\n");
+    } else {
+        printf("Some performance tests FAILED\n");
+    }
+    
+    return result;
 }
