+++ tests/test_platform_abstraction.c
@@ -88,47 +88,40 @@ static int test_universal_file_operations(void) {
       "BCPL Modernization Test Data - Universal Platform Support";
 
   // Test file creation and writing
-  bcpl_file_handle_t *file = bcpl_platform_fopen(test_filename, 'w', false);
+  FILE *file = bcpl_platform_fopen(test_filename, "w");
   if (!file) {
     printf("❌ Failed to create test file\n");
     return 0;
   }
 
-  // Write test data character by character
-  for (size_t i = 0; i < strlen(test_data); i++) {
-    if (bcpl_platform_fputc(test_data[i], file) < 0) {
-      printf("❌ Failed to write test data\n");
-      bcpl_platform_fclose(file);
-      return 0;
-    }
+  size_t written = fwrite(test_data, 1, strlen(test_data), file);
+  if (written != strlen(test_data)) {
+    printf("❌ Failed to write complete test data\n");
+    fclose(file);
+    return 0;
   }
 
-  bcpl_platform_fclose(file);
+  fclose(file);
   printf("✅ File creation and writing successful\n");
 
   // Test file reading
-  file = bcpl_platform_fopen(test_filename, 'r', false);
+  file = bcpl_platform_fopen(test_filename, "r");
   if (!file) {
     printf("❌ Failed to open test file for reading\n");
     return 0;
   }
 
   char read_buffer[256];
-  size_t read_bytes = 0;
-  int ch;
-  while ((ch = bcpl_platform_fgetc(file)) >= 0 &&
-         read_bytes < sizeof(read_buffer) - 1) {
-    read_buffer[read_bytes++] = (char)ch;
-  }
+  size_t read_bytes = fread(read_buffer, 1, sizeof(read_buffer) - 1, file);
   read_buffer[read_bytes] = '\0';
 
   if (strcmp(read_buffer, test_data) != 0) {
     printf("❌ Read data doesn't match written data\n");
-    bcpl_platform_fclose(file);
+    fclose(file);
     return 0;
   }
 
-  bcpl_platform_fclose(file);
+  fclose(file);
   printf("✅ File reading and data integrity verified\n");
 
   // Clean up
@@ -217,7 +210,7 @@ static int test_architecture_optimizations(void) {
 
   // Test optimized copy
   uint64_t start = bcpl_platform_get_time_ns();
-  bcpl_platform_memcpy(dst, src, test_size);
+  bcpl_platform_memcpy_optimized(dst, src, test_size);
   uint64_t end = bcpl_platform_get_time_ns();
 
   // Verify copy
@@ -283,7 +276,7 @@ static int test_thread_safety(void) {
 // MAIN TEST RUNNER
 // ============================================================================
 
-int test_platform_abstraction_suite(void) {
+int main(void) {
   printf("\n🚀 BCPL PLATFORM ABSTRACTION VALIDATION SUITE\n");
   printf("==============================================\n");
   printf("Testing complete elimination of platform-specific tech debt...\n\n");
