+++ tests/test_runtime.c
@@ -30,10 +30,6 @@
 // Include modernized runtime headers
 #include "universal_platform.h"
 
-// External test function declarations
-extern int test_platform_abstraction_suite(void);
-extern int test_memory_safety_suite(void);
-
 // ============================================================================
 // TEST FRAMEWORK INFRASTRUCTURE
 // ============================================================================
@@ -160,25 +156,24 @@ void test_platform_abstraction(void) {
   TEST_ASSERT(init_result == 0, "Platform initialization");
 
   // Test CPU feature detection
-  bcpl_cpu_features_t cpu_features = bcpl_platform_get_cpu_features();
+  uint32_t cpu_features = bcpl_platform_get_cpu_features();
   TEST_ASSERT(true,
               "CPU feature detection completed"); // Always passes, just logs
 
-  printf("    CPU Features detected: %s (%d cores)\n", cpu_features.arch_name,
-         cpu_features.core_count);
+  printf("    CPU Features detected: 0x%08X\n", cpu_features);
 
 #ifdef BCPL_ARCH_X86_64
-  if (cpu_features.feature_flags & BCPL_CPU_FEATURE_SSE2) {
+  if (cpu_features & BCPL_CPU_FEATURE_SSE2) {
     printf("    ✓ SSE2 support detected\n");
   }
-  if (cpu_features.feature_flags & BCPL_CPU_FEATURE_AVX) {
+  if (cpu_features & BCPL_CPU_FEATURE_AVX) {
     printf("    ✓ AVX support detected\n");
   }
-  if (cpu_features.feature_flags & BCPL_CPU_FEATURE_AVX2) {
+  if (cpu_features & BCPL_CPU_FEATURE_AVX2) {
     printf("    ✓ AVX2 support detected\n");
   }
 #elif defined(BCPL_ARCH_ARM64)
-  if (cpu_features.feature_flags & BCPL_CPU_FEATURE_NEON) {
+  if (cpu_features & BCPL_CPU_FEATURE_NEON) {
     printf("    ✓ NEON support detected\n");
   }
 #endif
@@ -336,13 +331,6 @@ int main(int argc, char **argv) {
   test_optimized_operations();
   test_error_handling();
   test_performance();
-
-  // Run additional test suites from other files
-  printf("\n🧪 Running Platform Abstraction Test Suite...\n");
-  test_platform_abstraction_suite();
-
-  printf("\n🧪 Running Memory Safety Test Suite...\n");
-  test_memory_safety_suite();
 
   // Print final results
   printf("\n📊 TEST RESULTS SUMMARY\n");
